micropatient : correction suite à la session de mentorat
