+++ manageAssesment/src/main/java/com/example/manageAssesment/controller/AssessController.java
@@ -4,9 +4,7 @@ import com.example.manageAssesment.model.Note;
 import com.example.manageAssesment.model.Patient;
 import com.example.manageAssesment.service.ConfDockerService;
 import com.example.manageAssesment.service.IAssesService;
-import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -17,7 +15,6 @@ import java.util.List;
 /**
  * Manage the requests linked to an Assessment
  */
-
 
 @Slf4j
 @RestController
