+++ manageNote/src/main/java/com/example/manageNote/repository/NoteRepository.java
@@ -1,7 +1,6 @@
 package com.example.manageNote.repository;
 
 import com.example.manageNote.model.Note;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
