+++ managePatient/managePatient/src/main/java/com/example/managePatient/controller/PatientController.java
@@ -11,21 +11,33 @@ import com.fasterxml.jackson.databind.cfg.CoercionInputShape;
 import com.fasterxml.jackson.databind.type.LogicalType;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import lombok.extern.slf4j.Slf4j;
+import org.json.JSONException;
+import org.json.JSONObject;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.ConstraintViolation;
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
 import java.util.List;
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Manage the requests linked to a Patient
  */
 @Slf4j
 @RestController
+
 public class PatientController {
     private final PatientRepository patientRepository;
     private final IPatientService patientService;
     private final ObjectMapper mapper;
+
+    private final Validator validator;
+
 
     public PatientController(PatientRepository patientRepository, IPatientService patientService) {
         this.patientRepository = patientRepository;
@@ -35,6 +47,9 @@ public class PatientController {
         mapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
         mapper.coercionConfigFor(LogicalType.Enum)
                 .setCoercion(CoercionInputShape.EmptyString, CoercionAction.AsNull);
+
+        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        validator = factory.getValidator();
     }
 
 
@@ -57,14 +72,12 @@ public class PatientController {
     public ResponseEntity<Patient> find(@PathVariable Integer id){
         Patient patientFromDB = patientService.findPatientById(id).orElse(null);
 
-        //If the patient doesn't exist, send status 404
         if(patientFromDB == null) {
-            //log.error("Error : id Patient doesn't exist in the Data Base.");
+            log.error("Error : id Patient doesn't exist in the Data Base.");
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
-        //log.info("Returning the patient's informations");
-        //else, return patient
+        log.info("Returning the patient's informations");
         return new ResponseEntity<>(patientFromDB, HttpStatus.OK);
     }
 
@@ -75,17 +88,14 @@ public class PatientController {
      */
     @GetMapping("/patient/exist/{id}")
     public ResponseEntity<Boolean> patientExist(@PathVariable Integer id){
-        //Récupère un patient dans la BD à partir de son ID. sinon renvoie null
         Patient patientFromDB = patientService.findPatientById(id).orElse(null);
 
-        //S'il n'éxiste pas, envoie statut 404
         if(patientFromDB == null) {
-            //log.error("Error : id Patient doesn't exist in the Data Base.");
+            log.error("Error : id Patient doesn't exist in the Data Base.");
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
-        //log.info("Returning the patient's informations");
-        //Sinon, retourner boolean et status de la requête
+        log.info("Returning the patient's informations");
         return new ResponseEntity<>(true, HttpStatus.OK);
     }
 
@@ -99,8 +109,7 @@ public class PatientController {
     public ResponseEntity<List<Patient>> patientsByGivenAndFamily(@RequestParam String given, @RequestParam String family){
         List<Patient> patients = patientService.findByGivenAndFamily(given, family);
 
-        //log.info("Returning the patient's informations");
-        //Sinon, retourner patient
+        log.info("Returning the patient's informations");
         return new ResponseEntity<>(patients, HttpStatus.OK);
     }
 
@@ -116,8 +125,7 @@ public class PatientController {
         if (patient == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        //log.info("Returning the patient's informations");
-        //Sinon, retourner patient
+        log.info("Returning the patient's informations");
         return new ResponseEntity<>(patient, HttpStatus.OK);
     }
 
@@ -130,13 +138,11 @@ public class PatientController {
     @PostMapping("/patient/add")
     public ResponseEntity<Object> addPatient(@RequestBody String body) throws JsonProcessingException {
 
-        //Convert request body to Json
         String patientJson = patientService.paramTojson(body);
-        //Convert json to Patient object
         Patient patient = mapper.readValue(patientJson, Patient.class);
 
-        // validation before save
-        ResponseEntity<Object> errorResponse = patientService.getValidationErrors(patient);
+        log.info("Validation before save");
+        ResponseEntity<Object> errorResponse = getValidationErrors(patient);
         if (errorResponse != null) {
             return errorResponse;
         }
@@ -154,15 +160,13 @@ public class PatientController {
     @PutMapping("/patient/{id}")
     public ResponseEntity<Object> update(@PathVariable Integer id, @RequestBody Patient patient) {
 
-        //Cherche person par son id
         Patient patientFromDB = patientService.findPatientById(id).get();
 
         if(patientFromDB == null) {
-            //log.error("Error : Patient already  exist in the Data Base.");
+            log.error("Error : Patient already  exist in the Data Base.");
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        //log.debug("Set properties in the object patientFromDB.");
-        //sinon mettre à jour person
+        log.debug("Set properties in the object patientFromDB.");
         patientFromDB.setAddress(patient.getAddress());
         patientFromDB.setGiven(patient.getGiven());
         patientFromDB.setFamily(patient.getFamily());
@@ -171,8 +175,8 @@ public class PatientController {
         patientFromDB.setDob(patient.getDob());
 
 
-        // validation before update
-        ResponseEntity<Object> errorResponse = patientService.getValidationErrors(patient);
+        log.info("validation before update");
+        ResponseEntity<Object> errorResponse = getValidationErrors(patient);
         if (errorResponse != null) {
             return errorResponse;
         }
@@ -199,4 +203,22 @@ public class PatientController {
     }
 
 
+
+    public ResponseEntity<Object> getValidationErrors(Patient patient) {
+        Set<ConstraintViolation<Patient>> violations = validator.validate(patient);
+        if (!violations.isEmpty()) {
+            AtomicInteger nb = new AtomicInteger(1);
+            JSONObject jsonError = new JSONObject();
+            violations.stream().forEach(v -> {
+                try {
+                    jsonError.put("Error " + nb.getAndIncrement(), v.getMessage());
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            });
+            log.error(jsonError.toString());
+            return new ResponseEntity<>(jsonError.toString(), HttpStatus.BAD_REQUEST);
+        }
+        return null;
+    }
 }
