+++ managePatient/managePatient/src/main/java/com/example/managePatient/service/PatientServiceImpl.java
@@ -2,35 +2,20 @@ package com.example.managePatient.service;
 
 import com.example.managePatient.model.Patient;
 import com.example.managePatient.repository.PatientRepository;
+import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import javax.validation.ConstraintViolation;
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 
+@AllArgsConstructor
 @Slf4j
 @Service
 public class PatientServiceImpl implements IPatientService{
     private final PatientRepository patientRepository;
-    private final Validator validator;
 
-    public PatientServiceImpl(PatientRepository patientRepository) {
-        this.patientRepository = patientRepository;
-
-        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
-        validator = factory.getValidator();
-    }
 
 
     @Override
@@ -72,7 +57,7 @@ public class PatientServiceImpl implements IPatientService{
     @Override
     public String paramTojson(String paramIn) {
         if (paramIn.startsWith("{")) {
-            //in this case it is already in json format
+            log.info("Param already in Json format");
             return paramIn;
         }
         paramIn = paramIn.replaceAll("=", "\":\"");
@@ -80,22 +65,5 @@ public class PatientServiceImpl implements IPatientService{
         return "{\"" + paramIn + "\"}";
     }
 
-    @Override
-    public ResponseEntity<Object> getValidationErrors(Patient patient) {
-        Set<ConstraintViolation<Patient>> violations = validator.validate(patient);
-        if (!violations.isEmpty()) {
-            AtomicInteger nb = new AtomicInteger(1);
-            JSONObject jsonError = new JSONObject();
-            violations.stream().forEach(v -> {
-                try {
-                    jsonError.put("Error " + nb.getAndIncrement(), v.getMessage());
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-            });
-            log.error(jsonError.toString());
-            return new ResponseEntity<>(jsonError.toString(), HttpStatus.BAD_REQUEST);
-        }
-        return null;
-    }
+
 }
